Refactor the common feature sheet functionalities
